+++ lib/rubocop/cop/style/optional_boolean_parameter.rb
@@ -4,7 +4,8 @@ module RuboCop
   module Cop
     module Style
       # This cop checks for places where keyword arguments can be used instead of
-      # boolean arguments when defining methods.
+      # boolean arguments when defining methods. `respond_to_missing?` method is allowed by default.
+      # These are customizable with `AllowedMethods` option.
       #
       # @example
       #   # bad
@@ -23,13 +24,20 @@ module RuboCop
       #     puts bar
       #   end
       #
+      # @example AllowedMethods: ['some_method']
+      #   # good
+      #   def some_method(bar = false)
+      #     puts bar
+      #   end
+      #
       class OptionalBooleanParameter < Base
+        include AllowedMethods
+
         MSG = 'Use keyword arguments when defining method with boolean argument.'
         BOOLEAN_TYPES = %i[true false].freeze
-        METHODS_EXCLUDED = %i[respond_to_missing?].freeze
 
         def on_def(node)
-          return if METHODS_EXCLUDED.include?(node.method_name)
+          return if allowed_method?(node.method_name)
 
           node.arguments.each do |arg|
             next unless arg.optarg_type?