+++ spec/rubocop/cop/style/optional_boolean_parameter_spec.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::Style::OptionalBooleanParameter do
-  subject(:cop) { described_class.new }
+RSpec.describe RuboCop::Cop::Style::OptionalBooleanParameter, :config do
+  let(:cop_config) do
+    { 'AllowedMethods' => [] }
+  end
 
   it 'registers an offense when defining method with optional boolean arg' do
     expect_offense(<<~RUBY)
@@ -49,10 +51,16 @@ RSpec.describe RuboCop::Cop::Style::OptionalBooleanParameter do
     RUBY
   end
 
-  it 'does not register an offense when defining respond_to_missing? method with boolean arg' do
-    expect_no_offenses(<<~RUBY)
-      def respond_to_missing?(arg, bar = false)
-      end
-    RUBY
+  context 'when AllowedMethods is not empty' do
+    let(:cop_config) do
+      { 'AllowedMethods' => %w[respond_to_missing?] }
+    end
+
+    it 'does not register an offense' do
+      expect_no_offenses(<<~RUBY)
+        def respond_to_missing?(method, include_all = false)
+        end
+      RUBY
+    end
   end
 end