+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 
 * [#8785](https://github.com/rubocop-hq/rubocop/pull/8785): Update TargetRubyVersion 2.8 to 3.0 (experimental). ([@em-gazelle][])
+* [#8650](https://github.com/rubocop-hq/rubocop/issues/8650): Faster find of hidden files in `TargetFinder` class which improves rubocop initial startup speed. ([@tleish][])
 
 ## 0.91.1 (2020-09-23)
 
@@ -4912,3 +4913,4 @@
 [@fsateler]: https://github.com/fsateler
 [@iSarCasm]: https://github.com/iSarCasm
 [@em-gazelle]: https://github.com/em-gazelle
+[@tleish]: https://github.com/tleish